PB-35 Move external asset tests to new file
PyLint critizied the length of the test module
